+++ src/models/Settings.php
@@ -17,7 +17,7 @@ class Settings extends Model
     public $silent = false;
     public $renderDelay = 0;
     public $keyboardShortcuts = true;
-    public $useNativeScreenshot = true;
+    public $useNativeScreenshot = false;
     public $extension = false;
 
     public function defineRules(): array
@@ -29,6 +29,6 @@ class Settings extends Model
 
     public function getProject(): string
     {
-        return App::parseEnv($this->project);
+        return strval(App::parseEnv($this->project));
     }
 }